azureml_pipeline/requirements.txt
# Base Personal Agendas dependencies (processors, LangChain stack, etc.)
-r ../requirements.txt

# Azure ML SDK + auth helpers
azure-ai-ml==1.20.0
azureml-fsspec==1.3.0
azureml-mlflow==1.59.0
mltable==1.6.1
azure-identity==1.17.1
azure-keyvault-secrets==4.9.0

# Observability stack (pinned to versions compatible with azure-monitor-opentelemetry 1.5.0)
azure-monitor-opentelemetry==1.5.0
azure-monitor-opentelemetry-exporter==1.0.0b26
azure-core-tracing-opentelemetry==1.0.0b11
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation==0.42b0
opentelemetry-instrumentation-django==0.42b0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-instrumentation-flask==0.42b0
opentelemetry-instrumentation-psycopg2==0.42b0
opentelemetry-instrumentation-requests==0.42b0
opentelemetry-instrumentation-urllib==0.42b0
opentelemetry-instrumentation-urllib3==0.42b0
opentelemetry-instrumentation-wsgi==0.42b0
opentelemetry-semantic-conventions==0.42b0
opentelemetry-util-http==0.42b0
opentelemetry-resource-detector-azure==0.1.4

# Runtime utilities
python-dotenv==1.0.1
mlflow-skinny
seaborn>=0.11.0

# Model runtimes frequently required by embeddings/recommendations
torch==2.5.1
